pyproject.toml
[tool.poetry]
name = "ordo-fast"
version = "0.1.0"
description = ""
authors = ["DevTarcizio <159829865+DevTarcizio@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.128.0"}
sqlalchemy = "^2.0.45"
pydantic-settings = "^2.12.0"
alembic = "^1.17.2"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
pyjwt = "^2.10.1"
tzdata = "^2025.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 82
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev ordo_fast/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=ordo_fast -vv'
post_test = 'coverage html'
commit = 'git commit -m'
post_commit = 'git push origin main'
create_migrations = 'alembic revision --autogenerate -m'
migrate = 'alembic upgrade'
demigrate = 'alembic downgrade'
shell_db = 'python -m sqlite3 database.db'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
